% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deserialise.R
\name{deserialise}
\alias{deserialise}
\title{Translate JSON-based simmer problem definition into a simmer environment}
\usage{
deserialise(json_obj)
}
\arguments{
\item{json_obj}{a string, URL or files}
}
\description{
Translate JSON-based simmer problem definition into a simmer environment
}

