% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deserialise.R
\name{process_value}
\alias{process_value}
\title{Process a value and translate to correct output}
\usage{
process_value(value)
}
\arguments{
\item{value}{value to process}
}
\description{
For internal usage.
}

